Compile with warnings enabled (#38)
* Compile on Windows with /W4 and fix all warnings

* Fix warnings for GCC

* Forward compiler and build type to srtp

* Remove unneeded namespace qualifier

* -Wextra and -Wall

+++ include/srtc/track.h
@@ -19,7 +19,7 @@ class Track
 {
 public:
     struct SimulcastLayer : srtc::SimulcastLayer {
-        uint16_t index = { 0 }; // [0..3]
+        uint8_t index = { 0 }; // [0..3]
     };
 
     struct CodecOptions {
